# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for IAD (Information and Analytical Data) system.",
        "title": "IAD Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/your-org/iad-backend",
            "email": "support@iad.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/stage-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of stage requests with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-requests"
                ],
                "summary": "Get stage requests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date from (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date to (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.StagesRequestsFilterResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/stage-requests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific stage request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-requests"
                ],
                "summary": "Get stage request by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.StageRequestDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing stage request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-requests"
                ],
                "summary": "Update stage request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stage request update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateStageRequestResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a stage request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-requests"
                ],
                "summary": "Delete stage request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/stage-requests/{id}/form": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Form a draft stage request into a submitted request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-requests"
                ],
                "summary": "Form stage request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/stage-requests/{id}/reject": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reject a stage request (moderator action)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-requests"
                ],
                "summary": "Reject stage request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/stage-requests/{id}/resolve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resolve a stage request (moderator action)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-requests"
                ],
                "summary": "Resolve stage request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.StageRequestDetailResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "stage_request_to_stages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.StageRequestToStageDetailResponse"
                    }
                }
            }
        },
        "handler.StageRequestToStageDetailResponse": {
            "type": "object",
            "properties": {
                "first_dimension_const": {
                    "type": "number"
                },
                "first_dimension_name": {
                    "type": "string"
                },
                "input_field_1": {
                    "type": "integer"
                },
                "input_field_2": {
                    "type": "integer"
                },
                "second_dimension_const": {
                    "type": "number"
                },
                "second_dimension_name": {
                    "type": "string"
                },
                "stage_title": {
                    "type": "string"
                }
            }
        },
...
        "handler.StagesRequestsFilterResponse": {
            "type": "object",
            "properties": {
                "ClosedAt": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "FormedAt": {
                    "type": "string"
                },
                "ModeratorID": {
                    "type": "integer"
                },
                "ProductName": {
                    "type": "string"
                },
                "Status": {
                    "type": "integer"
                },
                "UserID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateStageRequestResponse": {
            "type": "object",
            "properties": {
                "product_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}